requirements.txt
absl-py==1.2.0
astunparse==1.6.3
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.6.15.1
cffi @ file:///C:/Windows/Temp/abs_6808y9x40v/croots/recipe/cffi_1659598653989/work
charset-normalizer==2.1.0
click @ file:///C:/ci/click_1646056762388/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cryptography @ file:///C:/ci/cryptography_1652101729823/work
cycler==0.11.0
dill==0.3.5.1
docopt==0.6.2
editdistance==0.6.2
etils==0.8.0
filelock==3.8.0
Flask @ file:///opt/conda/conda-bld/flask_1648041541647/work
Flask-Cors==3.0.10
flatbuffers==1.12
fonttools==4.37.1
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.4
grpcio==1.47.0
h5py==3.7.0
heatmap==2.2.1
huggingface-hub==0.8.1
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imageio==2.22.0
importlib-resources==5.9.0
itsdangerous @ file:///tmp/build/80754af9/itsdangerous_1621432558163/work
Jinja2 @ file:///opt/conda/conda-bld/jinja2_1647436528585/work
keras==2.9.0
keras-nlp==0.3.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe @ file:///C:/ci_310/markupsafe_1642084045935/work
matplotlib==3.5.3
mkl-fft==1.3.1
mkl-random @ file:///C:/ci_310/mkl_random_1643050563308/work
mkl-service==2.4.0
networkx==2.8.6
numpy @ file:///C:/ci/numpy_and_numpy_base_1649782984482/work
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
Pillow==9.0.1
pipreqs==0.4.11
promise==2.3
protobuf==3.19.4
Py-BOBYQA==1.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclustering==0.10.1.2
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.4.1
pywer==0.1.1
PyYAML==6.0
regex==2022.7.25
requests @ file:///C:/ci/requests_1657735340829/work
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.9.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-datasets==4.6.0
tensorflow-estimator==2.9.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-metadata==1.10.0
tensorflow-text==2.9.0
termcolor==1.1.0
tifffile==2022.8.12
tokenizers==0.12.1
toml==0.10.2
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0
transformers==4.21.1
typing_extensions @ file:///C:/Windows/TEMP/abs_dd2d0moa85/croots/recipe/typing_extensions_1659638831135/work
urllib3 @ file:///C:/Windows/TEMP/abs_65ynz4fdmi/croots/recipe/urllib3_1659110473919/work
Werkzeug @ file:///opt/conda/conda-bld/werkzeug_1645628268370/work
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
wincertstore==0.2
wrapt==1.14.1
yarg==0.1.9
zipp==3.8.1
